requirements.txt
appdirs==1.4.4
astroid==2.4.2
autopep8==1.5.4
black==20.8b1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
filelock==3.0.12
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.19.4
pandas==1.1.4
pathspec==0.8.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
requests==2.25.0
requests-file==1.5.1
rope==0.18.0
schedule==0.6.0
six==1.15.0
tldextract==3.0.2
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wrapt==1.12.1
